Fixes #3.
In the tab panel, multiple elements were using the same ARIA ID. This
occurs because when elements of the panel get generated, they reuse the
same statically-defined ID. To fix this, replace the ID used with a
dynamically generated one. In this case, use the forloop counter to get
a unique number for each element in the panel, which also reflects the
naming behavior of some other elements in the side panel. After this
change, the Accessibility score reported in Lighthouse increased from 89
to 96.