Create initial db migrations for Reviews
Resolves #15.

Add initial script to create new database model to store reviews.
Rename all instances of "text" to "addlcomments" to be more descriptive
since more entries in the form are expected to be added.
Remove 'search.py' since it currently doesn't make sense to search
through the values that reviewers enter into the form.